+++ src/commands/command.ts
@@ -238,6 +238,12 @@ export class Command {
             component.setDisabled(true);
         }
     }
+    /**
+     * @param builder action row component containing page buttons
+     * @param allCondition if true, disable all buttons
+     * @param startCondition if true, disable first and previous
+     * @param endCondition if true, disable next and end
+     */
     protected disablePageButtons_check(builder: Discord.ActionRowBuilder<Discord.ButtonBuilder>, allCondition: boolean, startCondition: boolean, endCondition: boolean) {
         if (allCondition || (startCondition && endCondition)) {
             this.disableButtons(builder);