+++ src/tools/other.ts
@@ -309,12 +309,17 @@ export function userbitflagsToEmoji(flags: Discord.UserFlagsBitField) {
     return newArr;
 }
 
-export function scoreTotalHits(stats: osuapi.types_v2.ScoreStatistics) {
-    let total = 0;
-    for (const value in stats) {
-        total += stats[value];
+export function scoreTotalHits(stats: osuapi.types_v2.ScoreStatistics, ruleset: osuapi.Ruleset) {
+    switch (ruleset) {
+        case osuapi.Ruleset.osu: default:
+            return stats.great + (stats.ok ?? 0) + (stats.meh ?? 0) + (stats.miss ?? 0);
+        case osuapi.Ruleset.taiko:
+            return stats.great + (stats.good ?? 0) + (stats.miss ?? 0);
+        case osuapi.Ruleset.fruits:
+            return stats.great + (stats.ok ?? 0) + (stats.meh ?? 0) + stats.small_tick_hit + (stats.miss ?? 0);
+        case osuapi.Ruleset.mania:
+            return (stats.perfect ?? 0) + stats.great + stats.good + (stats.ok ?? 0) + (stats.meh ?? 0) + (stats.miss ?? 0);
     }
-    return total;
 }
 
 export function scoreIsComplete(
@@ -322,8 +327,9 @@ export function scoreIsComplete(
     circles: number,
     sliders: number,
     spinners: number,
+    ruleset: osuapi.Ruleset
 ) {
-    let total = scoreTotalHits(stats);
+    let total = scoreTotalHits(stats, ruleset);
     return {
         passed: total == circles + sliders + spinners,
         objectsHit: total,