+++ changelog.md
@@ -6,6 +6,17 @@
 
 [commit](https://github.com/sbrstrkkdwmdr/ssob)</br>
 
+### Fixed
+
+-   (docs) `ScoreListCommand` - missing star rating option for score sorting
+-   (docs) "Formatted as" being on the same line as arg description
+-   (docs) `ScoreListCommand` - sort by default is now recent on some lists
+
+### Changed
+
+-   (docs) show arg options
+-   (docs) "default" values are now encased in code blocks
+
 ## [5.5.0] - 2025-11-01
 
 [commit](https://github.com/sbrstrkkdwmdr/ssob/commit/bfa78bbf0f23ef903fcceb9cc9fd12810c296154)</br>